% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_funcs.R
\name{mult_abund_norm}
\alias{mult_abund_norm}
\title{Perform RJMCMC for posterior sampling of multivariate cluster abundance model with Gaussian 
observations}
\usage{
mult_abund_norm(data_list, initial_list, prior_list, block, begin_group_update,
  update_omega = T, burn, iter)
}
\arguments{
\item{data_list}{A named list created of data items created from user data with the 
function \code{multAbund::make_data_list}. This data will be used for model fitting.}

\item{initial_list}{A named list of initial values for the parameters (see details). 
The functions \code{multAbund::sugs} or \code{multAbund::make_inits} can be used to create this
list.}

\item{prior_list}{A named list of prior distribution parameters. See details.}

\item{block}{Number of iterations between Metropolis proposal adaptation.}

\item{begin_group_update}{The iteration at with the group clusters begin updating.
The RJMCMC often performs better when the chain is allowed to sample only the parameters 
before the groups begin updating.}

\item{update_omega}{Logical. Should omega be updated or not?}

\item{burn}{Number of burnin iterations that are discarded.}

\item{iter}{Number of iterations retained for posterior inference.}
}
\description{
Fit a Dirichlet Process random effect model for joint species distribution inference of Gaussian abundance 
(typically resulting from a log transformation) data using a RJMCMC procedure. THIS IS CURRENTLY A
TEST FUNCTION THAT IS NOT FULLY DEVELOPED YET! IT USES A DIFFERENT VARIANCE MODEL FOR DELTA PARAMETERS.
}
\details{
The \code{prior_list} argument needs to contain the following items:
\itemize{
\item a_alpha numeric. shape parameter for gamma prior on alpha,
\item b_alpha numeric. scale parameter for gamma prior on alpha,
\item Sigma_beta_inv numeric matrix. precision matrix for global regression coefficients, beta
\item mu_beta numeric vector. prior mean of beta parameters
\item phi_omega numeric. scale for half t/normal prior on delta variance parameter (omega),
\item df_omega numeric. degrees of freedom for half-t prior on omega (df_omega>=50 means a half-normal will be assumed)
\item phi_sigma numeric. scale parameter for half-t/normal prior for sigma parameters
\item df_sigma numeric. degrees of freedom for sigma prior (>=50 implies half-normal will be used)
}
}
\section{Vignette}{

A demonstration using simulated data is available in vignette form and can be accessed 
by typing \verb{vignette("simulation_demo", package="multAbund")}. The vignette demo 
uses a realistic number of iterations, so, if the user decides to run the associated \verb{R}
code it will take some time.
}

\author{
Devin S. Johnson
}
