% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multAbund-package.R
\docType{package}
\name{multAbund-package}
\alias{multAbund-package}
\alias{multAbund}
\title{Use Dirichlet Process Prior to Model Multivariate Animal Abundance and Occurance}
\description{
Functions for implementing the RJMCMC estimation of multivariate abundence and occurence. 
These functions model multivariate abundence with latent group 
designations for each species. The associations between species are 
functions of common group membership.

\tabular{ll}{ 
Package: \tab multAbund\cr 
Type: \tab Package\cr 
Version: \tab 0.03.9001\cr 
Date: \tab March 28, 2018\cr 
License: \tab CC0 \cr 
LazyLoad: \tab yes\cr 
}
}
\author{
Devin S. Johnson

Maintainer: Devin S. Johnson <devin.johnson@noaa.gov>
}
