% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_funcs.R
\name{pois_reg}
\alias{pois_reg}
\title{Probit regression via MCMC}
\usage{
pois_reg(data_list, pred_list, phi_beta, mu_beta, phi_sigma, df_sigma, block,
  burn, iter)
}
\arguments{
\item{data_list}{A named list created of data items created from user data with the 
function \code{multAbund::make_data_list}. This data will be used for model fitting.}

\item{pred_list}{A named list created in the same way as \code{data_list}, however, 
this list will be used for prediction purposes. For example, a user may want to withold
data in a cross-validation check.}

\item{phi_beta}{Scale parameter for the prior covariance of \verb{beta}. The covariance 
is defined to be \verb{phi_beta^2*solve(X'X)}, where \code{X} is the design matrix 
for the covariates associated with \verb{beta}}

\item{mu_beta}{Prior mean of \verb{beta}.}

\item{phi_sigma}{Scale parameter for the half-t prior distribution for \verb{sigma}.}

\item{df_sigma}{Degrees of freedom for \verb{sigma} prior. Follows the same rules as the omega prior specification.}

\item{block}{Number of iterations between Metropolis proposal adaptation.}

\item{burn}{Number of burnin iterations that are discarded.}

\item{iter}{Number of iterations retained for posterior inference.}
}
\description{
Fit a Poisson regression model with normal randome effect for modeling overdispersion.
}
\author{
Devin S. Johnson
}

