% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction_sample}
\alias{prediction_sample}
\title{Predictions for Dirichlet Process Joint Species Distribution Model}
\usage{
prediction_sample(object, pred_list)
}
\arguments{
\item{object}{A fitted MCMC model object obtained from call to one of the \code{multAbund}
fitting functions.}

\item{pred_list}{A list similar to data_list in the MCMC fitting functions 
except values are used where prediction is desired}

\item{z_pred}{Alogical indication of whether the posterior of z should 
be used (\code{z_pred=FALSE}) or the predictive distribution for z, given the
parameter posterior distribution (\code{z_pred=TRUE}; the default). It is ignored
for probit occurence models}
}
\description{
Create MCMC sample from various forms of the predictive distrubution 
using the MCMC sample of the parameters
}

