% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_funcs.R
\name{probit_reg}
\alias{probit_reg}
\title{Probit regression via MCMC}
\usage{
probit_reg(data_list, pred_list, burn, iter)
}
\arguments{
\item{data_list}{A named list created of data items created from user data with the 
function \code{multAbund::make_data_list}. This data will be used for model fitting.}

\item{pred_list}{A named list created in the same way as \code{data_list}, however, 
this list will be used for prediction purposes. For example, a user may want to withold
data in a cross-validation check.}

\item{burn}{Number of burnin iterations that are discarded.}

\item{iter}{Number of iterations retained for posterior inference.}

\item{prior_list}{A list of prior parameters with the names and elements: 
(1) \verb{Sigma_beta_inv}, the inverse covariance matrix of the normal prior for beta, and 
(2) \verb{mu_beta}, the prior mean of beta.}
}
\description{
Fit a probit regression model via MCMC using data augmentation.
}
\author{
Devin S. Johnson
}

