% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sugs.R
\name{sugs}
\alias{sugs}
\title{Compute intial values for the abundance / occurence RJMCMC using an approximation of the SUGS algorithm}
\usage{
sugs(design_matrices, log_alpha = 0, phi_beta = 10, mu_beta,
  phi_omega = 1, df_omega = 1)
}
\arguments{
\item{design_matrices}{A list created by the make_data_list function.}

\item{log_alpha}{An intial guess at the log of the Dirichlet process parameter. Defaults to 0.}

\item{phi_beta}{The prior variance parameter for the Normal prior for the fixed effect beta parameters}

\item{mu_beta}{The prior mean for beta}

\item{phi_omega}{The scale parameter for the half-t prior distribution for omega}

\item{df_omega}{Degrees-of-freedom parmeter for omega half-t prior distribution.}
}
\value{
A list with the following elements:
\item{beta}{A vector with the initial beta value}
\item{delta}{A vector with the initial delta values}
\item{groups}{Initial group assignments for each species}
\item{omega}{Inital value for omega}
\item{sigma}{Vector of inital values for sigma. Currently this is a vector filled with exp(-10). i.e., very little overdispersion.}
\item{log_alpha}{Initial value of log alpha for the Dirichlet process prior}
}
\description{
Compute intial values for the abundance / occurence RJMCMC using an approximation of the SUGS algorithm
}
\author{
Devin S. Johnson
}
\references{
Wang, L. and Dunson, D.B. (2011) Fast Bayesian Inference in Dirichlet Process Mixture Models. Journal of Computational and Graphical Statistics, 20:196-216.
}

