% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/view_example.R
\name{view_example}
\alias{view_example}
\title{View simulation example}
\usage{
view_example()
}
\description{
Users can view an .html file that contains an example with simulated abundance data.
The demo is similar to a vigette but it takes much too long to build.
}
\author{
Devin S. Johnson
}

