% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_opt_alpha_prior.R
\name{get_opt_alpha_prior}
\alias{get_opt_alpha_prior}
\title{Obtain parameters for gamma prior such that the distribution of the number of groups is equal to the \code{target} argument}
\usage{
get_opt_alpha_prior(n, target = NULL)
}
\arguments{
\item{n}{An integer representing the maximum number of groups. Typically this is the number of individuals.}

\item{target}{The desired prior probability mass function for the number of groups. This function must return a
vector of length \code{n} which contains the prior probabilities for each group size.}
}
\description{
This function uses the method of Dorazio (2009) to calculate the parameters of a gamma prior on the Dirichlet process such that the induced distribution function
of the number of clusters is equal to 1/k.
}
\author{
Devin S. Johnson
}
\references{
Dorazio, R. M. (2009) On selecting a prior for the precision parameter of Dirichlet process mixture models. Journal of Statistical Planning and Inference 139:3384-3390.
}

