% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_design.R
\name{make_data_list}
\alias{make_data_list}
\title{Make necessary data matrices for multivariate abundance and occurence RJMCMC functions}
\usage{
make_data_list(counts = ~1, delta_model, X_additional, sigma_model = ~1,
  data)
}
\arguments{
\item{counts}{A formula that provides the column which contains the counts. Must be of the form \code{~ var - 1} to avoid having a extra column of 1s.}

\item{delta_model}{A formula decribing the variables for which delta will be the coefficients.}

\item{X_additional}{A formula giveng additional variables to be used for all groups.}

\item{sigma_model}{A formula describing the sigma vector. Must be of the form \code{~factorVariable-1}.}

\item{data}{Data.frame containing counts and environmental variables}
}
\value{
A list containing the following elements:
\item{n}{A vector of counts}
\item{H}{The environmental variables from which the clustering is based}
\item{X}{The design matrix for the global covariates}
\item{D}{Design matrix defining sigma_ij}
\item{data}{Data frame containing the original data}
}
\description{
Make necessary data matrices for multivariate abundance and occurence RJMCMC functions
}
\author{
Devin S. Johnson
}

