% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design.R
\name{make_data_list}
\alias{make_data_list}
\title{Make necessary data matrices for multivariate abundance and occurence RJMCMC functions}
\usage{
make_data_list(
  counts = NULL,
  occur = NULL,
  z_col = NULL,
  delta_model,
  X_model,
  sigma_model = ~1,
  gamma_model = ~1,
  data
)
}
\arguments{
\item{counts}{A character string providing the name of the column in \verb{data} that contains the counts}

\item{occur}{A character string providing the name of the column in \verb{data} that contains the binary occurence data}

\item{z_col}{A character string providing the name of the column in \verb{data} that contains the continuous abundance data}

\item{delta_model}{A formula decribing the variables for which delta will be the coefficients.}

\item{X_model}{A formula giveng additional variables to be used for all groups.}

\item{sigma_model}{A formula describing the sigma vector. Must be of the form \verb{~factorVariable-1}.}

\item{gamma_model}{A formula describing the gamma vector for ZIP models}

\item{data}{Data.frame containing counts and environmental variables}
}
\value{
A list containing the following elements:
\item{counts}{The column containing count data}
\item{occur}{The column containing occurence data}
\item{z_col}{The column containing Gaussian response data}
\item{H}{The environmental variables from which the clustering is based}
\item{X}{The design matrix for the global covariates}
\item{D}{Design matrix defining sigma_ij}
\item{M}{Design matrix defining the gamma_ij}
\item{data}{Data frame containing the original data}
}
\description{
Makes the necessary lists for imput into the RJMCMC fitting functions
}
\author{
Devin S. Johnson
}
