% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_funcs.R
\name{mult_abund_norm}
\alias{mult_abund_norm}
\title{Perform RJMCMC for posterior sampling of multivariate cluster abundance model with Gaussian 
observations}
\usage{
mult_abund_norm(data_list, initial_list, prior_list, block, begin_group_update,
  max_adapt, update_omega = T, burn, iter)
}
\arguments{
\item{data_list}{A named list created of data items created from user data with the 
function \code{multAbund::make_data_list}. This data will be used for model fitting.}

\item{initial_list}{A named list of initial values for the parameters (see details). 
The functions \code{multAbund::sugs} or \code{multAbund::make_inits} can be used to create this
list.}

\item{block}{Number of iterations between Metropolis proposal adaptation.}

\item{begin_group_update}{The iteration at with the group clusters begin updating.
The RJMCMC often performs better when the chain is allowed to sample only the parameters 
before the groups begin updating.}

\item{max_adapt}{maximum number of iterations that proposals are adapted}

\item{update_omega}{Logical. Should omega be updated or not?}

\item{burn}{Number of burnin iterations that are discarded.}

\item{iter}{Number of iterations retained for posterior inference.}

\item{phi_beta}{Scale parameter for the prior covariance of \verb{beta}. The covariance 
is defined to be \verb{phi_beta^2*solve(X'X)}, where \code{X} is the design matrix 
for the covariates associated with \verb{beta}}

\item{mu_beta}{Prior mean of \verb{beta}.}

\item{phi_omega}{Scale parameter for omega prior. The omega prior is a half-t
distribution.}

\item{df_omega}{Degrees of freedom for the omega prior distribution. 
If \verb{df_omega > 50} then the prior is set to a half-normal distribution rather
than a half-t. If \verb{df_omega = 1}, then a half-Cauchy distribution is used.}

\item{a_alpha}{Shape parameter for the gamma distribution of the \verb{alpha} parameter.}

\item{b_alpha}{Rate parameter for the gamma distribution of the \verb{alpha} parameter.}

\item{phi_sigma}{Scale parameter for the half-t prior distribution for \verb{sigma}.}

\item{df_sigma}{Degrees of freedom for \verb{sigma} prior. Follows the same rules as the omega prior specification.}
}
\description{
Fit a Dirichlet Process random effect model for joint species distribution inference of Gaussian abundance 
(typically resulting from a log transformation) data using a RJMCMC procedure.
}
\details{
Here are some details.
}
\section{Vignette}{

A demonstration using simulated data is available in vignette form and can be accessed 
by typing \verb{vignette("simulation_demo", package="multAbund")}. The vignette demo 
uses a realistic number of iterations, so, if the user decides to run the associated \verb{R}
code it will take some time.
}

\author{
Devin S. Johnson
}
