% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_funcs.R
\name{pois_reg}
\alias{pois_reg}
\title{Probit regression via MCMC}
\usage{
pois_reg(data_list, pred_list, prior_list, initial_list, block, burn, iter)
}
\arguments{
\item{data_list}{A named list created of data items created from user data with the 
function \code{multAbund::make_data_list}. This data will be used for model fitting.}

\item{pred_list}{A named list created in the same way as \code{data_list}, however, 
this list will be used for prediction purposes. For example, a user may want to withold
data in a cross-validation check.}

\item{prior_list}{A list of prior parameters with the names and elements: 
(1) \verb{Sigma_beta_inv}, the inverse covariance matrix of the normal prior for beta,
(2) \verb{mu_beta}, the prior mean of beta, 
(3) \verb{phi_sigma} scale parameter for the half-t prior distribution for sigma,
(4) \verb{df_sigma} the degrees of freedom for sigma prior. Follows the same rules as the omega prior specification.}

\item{block}{Number of iterations between Metropolis proposal adaptation.}

\item{burn}{Number of burnin iterations that are discarded.}

\item{iter}{Number of iterations retained for posterior inference.}
}
\description{
Fit a Poisson regression model with normal randome effect for modeling overdispersion.
}
\author{
Devin S. Johnson
}

